+++ CHANGELOG.md
@@ -7,7 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### New
 * Created base project structure with elements taken from the previous Papaya project.
-
+* Added file input, classification, basic scoring, and file output.
+* Added sample parquet files for local development
 
 ### Changes